requirements.txt
# Dependencies for the framework
docopt
configparser
PyYAML
